Let liblights adjust the brightness of LEDs while previewing it
Each device might use the information about the brightness of LEDs
in a unique way. Currently the slider to adjust the brightness of
LEDs is using an already scaled RGB color to preview the intensity
of the light. Instead of doing this, take the default color of LED
notifications and add to it the information about the brightness.
This will let liblights adjust the brightness.

Change-Id: I4dd35944debc744ea6c7f1fcc5dd7a909e8863fa